Move maps functionality to MuseumMapFragment
