+++ app/src/main/java/nl/biancavanschaik/android/museumkaart/util/PermissionUtils.kt
@@ -1,6 +1,5 @@
 package nl.biancavanschaik.android.museumkaart.util
 
-import android.Manifest
 import android.app.Activity
 import android.content.Context
 import android.content.pm.PackageManager
@@ -12,5 +11,5 @@ fun Context.hasPermission(permission: String) =
         ContextCompat.checkSelfPermission(this, permission) == PackageManager.PERMISSION_GRANTED
 
 fun Activity.requestPermission(permission: String, @IntRange(from = 0) requestCode: Int) {
-    ActivityCompat.requestPermissions(this, arrayOf(Manifest.permission.ACCESS_FINE_LOCATION), requestCode)
+    ActivityCompat.requestPermissions(this, arrayOf(permission), requestCode)
 }
